requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws-lambda>=1.132.0
aws-cdk.aws-apigateway>=1.132.0
aws-cdk.aws-codepipeline>=1.132.0
aws-cdk.aws-codepipeline_actions>=1.132.0
aws-cdk.pipelines>=1.132.0
